gpu: intel: jit: ir: eltwise: change switch type to fix the warning

+++ src/gpu/intel/jit/ir/eltwise.hpp
@@ -44,7 +44,7 @@ public:
     }
 
     std::string str() const override {
-        switch (alg_kind) {
+        switch (static_cast<int>(alg_kind)) {
             case alg_kind::eltwise_relu: return "relu";
             case alg_kind::eltwise_tanh: return "tanh";
             case alg_kind::eltwise_elu: return "elu";
@@ -79,6 +79,8 @@ public:
             case alg_kind::eltwise_clip_v2_use_dst_for_bwd:
                 return "clip_v2_use_dst_for_bwd";
             case alg_kind::eltwise_round: return "round";
+            // Note: `eltwise_stochastic_round` is not a part of `enum` which
+            // forces `switch` to iterate over `int`, not `alg_kind_t`.
             case alg_kind::eltwise_stochastic_round: return "stochastic_round";
             default: ir_error_not_expected();
         }
